nus_fintech_gang_plus_side_characters/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
base58==2.1.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Deprecated==1.3.1
ECPy==1.2.5
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
starlette==0.50.0
types-Deprecated==1.3.1.20251101
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
wrapt==2.0.1
xrpl-py==4.4.0

